Fix all 19 engine clippy warnings; consume event payloads in world example
First clippy pass ever run against the engine (its CI never executed
while the source sat inside the zip). All fixes verified behavior-
preserving by the engine's 188 tests plus a headless run of the world
example.

- gamepad.rs, platformer.rs: map_or simplifications (clippy --fix),
  re-formatted to rustfmt style afterward
- ecs.rs: removed Column::contains -- dead code on a private trait,
  never called anywhere
- rng.rs tests: manual `lo <= v && v < hi` checks replaced with
  Range::contains
- tween.rs test: `fires` was inferred i32 and then cast to i32; typed
  the binding instead
- roguelike.rs: seed literal 0xDEAD_CA7 regrouped as 0x0DEA_DCA7 --
  identical value, equal-size digit groups
- world.rs: the example drained its event bus but discarded every
  payload (the two "field never read" warnings were pointing at a real
  demo gap). The drain loop now folds payloads into session state and
  the final summary prints them, so the example actually demonstrates
  event consumption:
  "Session complete -- steps=244 goal=true last_move=(873,60) sounds=[coin]"

cargo clippy --workspace --all-targets: 0 warnings.
cargo fmt --all --check: clean.

+++ izanagi/src/tween.rs
@@ -194,13 +194,13 @@ mod tests {
     #[test]
     fn timer_every_fires_repeatedly() {
         let mut timer = Timer::every(0.1);
-        let mut fires = 0;
+        let mut fires = 0i32;
         for _ in 0..100 {
             if timer.tick(0.02) {
                 fires += 1;
             }
         }
-        assert!((fires as i32 - 20).abs() <= 1);
+        assert!((fires - 20).abs() <= 1);
     }
 
     #[test]
